pyproject.toml
[project]
name = "docuflex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "python-multipart>=0.0.9",
    "jinja2>=3.1.4",
    "pymupdf>=1.24.14",
    "pillow>=10.4.0",
    "python-docx>=1.1.2",
    "pytesseract>=0.3.13",
    "pdf2docx>=0.5.8",
    "pypdf>=4.3.1",
]
